Start to implement --pretend option

+++ orchestra/cmds/clean.py
@@ -20,8 +20,13 @@ def handle_clean(args, config: Configuration):
         exit(1)
 
     build_dir = build.install.environment["BUILD_DIR"]
-    shutil.rmtree(build_dir, ignore_errors=True)
+    logger.info(f"Cleaning build dir for {build.qualified_name} ({build_dir})")
+
+    if not args.pretend:
+        shutil.rmtree(build_dir, ignore_errors=True)
 
     if args.include_sources:
         sources_dir = build.install.environment["SOURCE_DIR"]
-        shutil.rmtree(sources_dir, ignore_errors=True)
+        logger.info(f"Cleaning source dir for {build.qualified_name} ({sources_dir})")
+        if not args.pretend:
+            shutil.rmtree(sources_dir, ignore_errors=True)
